backend/requirements.txt
django
djangorestframework
django-filter
mysqlclient
numpy
pandas